requirements.txt
pandas==2.2.2
yfinance==0.2.52
requests==2.32.3
jinja2==3.1.4
matplotlib==3.9.0
openpyxl==3.1.5
python-dateutil==2.9.0.post0
pandas-datareader==0.10.0
PyYAML==6.0.1
